+++ packages/breakpoint/CHANGELOG.md
@@ -1,7 +1,11 @@
+## [4.1.0] 2019-12-04
+### Changed
+- Can now get name of current breakpoint in JS using getBreakpoint() or useBreakpoint()
+
 ## [4.0.0] 2019-9-23 [Release notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
 
 ## [3.0.0] 2019-9-23 [Release notes](https://creditandfinance.atlassian.net/wiki/spaces/HUB/pages/803930391/Upcoming+Major+Changes)
-### Updated 
+### Updated
 - Update typescript to version 3.5.2 and react to version 16.8.4
 
 ## [2.1.0] - 2019-05-17 - [HUB-174](https://creditandfinance.atlassian.net/browse/HUB-174)
@@ -14,7 +18,7 @@
 ### Changed
 - Updated CHANGELOG to match required parsing format.
 
-## [2.0.1] - 2019-05-08 - [HUB-54](https://creditandfinance.atlassian.net/browse/HUB-54) 
+## [2.0.1] - 2019-05-08 - [HUB-54](https://creditandfinance.atlassian.net/browse/HUB-54)
 ### Changed
 - Change module main entry point to esm in package.json